+++ src/App.js
@@ -5,83 +5,21 @@ import SetTracker from './SetTracker';
 import {useStickyState} from './hooks/useStickyState';
 import Hamburger from './Hamburger';
 import {DialogModal} from "./utils/DialogModal";
+import SettingsDialog from "./SettingsDialog";
 
 function App() {
   const [trackerType, setTrackerType] = useStickyState(1, 'countTracker.trackerType');
 
   const [setsCount, setSetsCount] = useStickyState(10, 'countTracker.setsCount');
   const [secondsInASet, setSecondsInASet] = useStickyState(20, 'countTracker.secondsInASet');
+  const [startSetDelay, setStartSetDelay] = useStickyState(2.5, 'countTracker.startSetDelay');
   const [mute, setMute] = useStickyState(false, 'countTracker.mute');
   const [isSettingsDialogOpened, setIsSettingsDialogOpened] = useState(false);
 
-  function toggleMute() {
-      setMute(mute => !mute);
-  }
-
-  function secondsInASetChanged(e) {
-      setSecondsInASet(parseInt(e.target.value));
-  }
-
-  function setsCountChanged(e) {
-      setSetsCount(parseInt(e.target.value));
-  }
-
-  const trackerTypChanged = (e) => {
-    setTrackerType(parseInt(e.target.value));
-  };
-
+  
   function onSettings() {
     setIsSettingsDialogOpened(true);
   }
-
-  const settingDialogContent = (
-    <div>
-        <label>Mute<input type="checkbox" name="mute" onChange={toggleMute} checked={mute}/></label>
-        <div>
-        <label>Seconds per set:
-            <select value={secondsInASet.toString()} onChange={secondsInASetChanged}>
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-                <option value="5">5</option>
-                <option value="6">6</option>
-                <option value="7">7</option>
-                <option value="8">8</option>
-                <option value="9">9</option>
-                <option value="10">10</option>
-                <option value="15">15</option>
-                <option value="20">20</option>
-                <option value="30">30</option>
-            </select>
-        </label>
-        </div>
-        <div>
-        <label>Sets Count:
-            <select value={setsCount.toString()} onChange={setsCountChanged}>
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-                <option value="5">5</option>
-                <option value="6">6</option>
-                <option value="7">7</option>
-                <option value="8">8</option>
-                <option value="9">9</option>
-                <option value="10">10</option>
-            </select>
-        </label>
-        </div>
-        <div>
-        <label>View:
-            <select value={trackerType.toString()} onChange={trackerTypChanged}>
-                <option value="1">Count Tracker</option>
-                <option value="2">Set Tracker</option>
-            </select>
-        </label>
-        </div>
-    </div>
-  );
 
   return (
 
@@ -92,7 +30,13 @@ function App() {
           isOpened={isSettingsDialogOpened}
           onProceed={null}
           onClose={() => setIsSettingsDialogOpened(false)}>
-          {settingDialogContent}
+          <SettingsDialog
+            mute={mute} setMute={setMute}
+            secondsInASet={secondsInASet} setSecondsInASet={setSecondsInASet}
+            setsCount={setsCount} setSetsCount={setSetsCount}
+            startSetDelay={startSetDelay} setStartSetDelay={setStartSetDelay}
+            trackerType={trackerType} setTrackerType={setTrackerType}
+          />
       </DialogModal>
 
       <div className='title-continer'>
@@ -102,7 +46,7 @@ function App() {
       </div>
 
       {trackerType === 1 ? <CountTracker mute={mute} /> :
-        <SetTracker setsCount={setsCount} secondsInASet={secondsInASet} mute={mute} />}
+        <SetTracker setsCount={setsCount} secondsInASet={secondsInASet} startSetDelay={startSetDelay} mute={mute} />}
 
     </div>
   );